requirements.txt
deepface==0.0.93
paddlepaddle>=2.6.2
paddleocr>=2.10.0
torch=2.2.0
psutil
PyYAML
torchvision
#psycopg[pool,binary]>=3.1.10
#pgvector
numpy==1.26.4
pyyaml>=6.0.1
#fastapi
#fastapi[standard]
google-api-python-client
google-auth
tf-keras
pillow-heif